#1883 Moved test_add_lfric_precision_symbol to lfric_symbol_table_test.
